[tools] perf/probe: Support distro-style debuginfo for uprobe
Message-id: <1410181752-25631-74-git-send-email-jolsa@redhat.com>
Patchwork-id: 91851
O-Subject: [PATCH RHEL7.1 BZ1134356 073/307] perf probe: Support distro-style debuginfo for uprobe
Bugzilla: 1134356
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1134356
https://bugzilla.redhat.com/show_bug.cgi?id=1134356

upstream
========
commit a15ad2f5360c821f030c53266ebf467738249c68
Author: Masami Hiramatsu <masami.hiramatsu.pt@hitachi.com>
Date: Thu Feb 6 05:32:27 2014 +0000

description
===========
Support distro-style debuginfo supported by dso for setting uprobes.
Note that this tries to find a debuginfo file based on the real path of
the target binary. If the debuginfo is not correctly installed on the
system, this can not find it.
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>