[tools] perf/symbols: Record the reason for filtering an address_location
Message-id: <1410181752-25631-106-git-send-email-jolsa@redhat.com>
Patchwork-id: 91756
O-Subject: [PATCH RHEL7.1 BZ1134356 105/307] perf symbols: Record the reason for filtering an address_location
Bugzilla: 1134356
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1134356
https://bugzilla.redhat.com/show_bug.cgi?id=1134356

upstream
========
commit b3cef7f60f17d953545f7069f6407fc24202a64d
Author: Namhyung Kim <namhyung@kernel.org>
Date: Mon Mar 17 16:59:21 2014 -0300

description
===========
By turning the addr_location->filtered member from a boolean to a u8
bitmap, reusing (and extending) the hist_filter enum for that.

This patch doesn't change the logic at all, as it keeps the meaning of
al->filtered !0 to mean that the entry _was_ filtered, so no change in
how this value is interpreted needs to be done at this point.

This will be soon used in upcoming patches.

[ yanked this out of a previous patch ]
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>