[tools] perf/symbols: Apply all filters to an addr_location
Message-id: <1410181752-25631-107-git-send-email-jolsa@redhat.com>
Patchwork-id: 91857
O-Subject: [PATCH RHEL7.1 BZ1134356 106/307] perf symbols: Apply all filters to an addr_location
Bugzilla: 1134356
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1134356
https://bugzilla.redhat.com/show_bug.cgi?id=1134356

upstream
========
commit 466fa7647413665dcba46c4f7f2b4a9808426989
Author: Namhyung Kim <namhyung@kernel.org>
Date: Mon Mar 17 17:12:27 2014 -0300

description
===========
Instead of bailing out as soon as we find a filter that applies, go on
checking all of them so that we can zoom in/out filters.

We also need to make sure we only update al->filtered after
thread__find_addr_map(), because there is where al->filtered gets
initialized to zero.

This will increase the cost of processing when all we don't need this
toggling, but will provide flexibility for the TUI and GTK+ interfaces,
that will incur in creating the hist_entries just once.

[ yanked this out of a previous patch ]
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>