[tools] perf: Allow ability to map cpus to nodes easily
Message-id: <1410181752-25631-159-git-send-email-jolsa@redhat.com>
Patchwork-id: 91824
O-Subject: [PATCH RHEL7.1 BZ1134356 158/307] perf tools: Allow ability to map cpus to nodes easily
Bugzilla: 1134356
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1134356
https://bugzilla.redhat.com/show_bug.cgi?id=1134356

upstream
========
commit 7780c25bae59fd042281710f5e1243268e2c18dc
Author: Don Zickus <dzickus@redhat.com>
Date: Mon Apr 7 14:55:21 2014 -0400

description
===========
This patch figures out the max number of cpus and nodes that are on the
system and creates a map of cpu to node.  This allows us to provide a cpu
and quickly get the node associated with it.

It was mostly copied from builtin-kmem.c and tweaked slightly to use less memory
(use possible cpus instead of max).  It also calculates the max number of nodes.

[ Removing out label code in init_cpunode_map ]
[ Adding check for snprintf error ]
[ Removing unneeded returns ]
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>