[tools] perf: Update cpumode for each cumulative entry
Message-id: <1410181752-25631-240-git-send-email-jolsa@redhat.com>
Patchwork-id: 91951
O-Subject: [PATCH RHEL7.1 BZ1134356 239/307] perf tools: Update cpumode for each cumulative entry
Bugzilla: 1134356
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1134356
https://bugzilla.redhat.com/show_bug.cgi?id=1134356

upstream
========
commit c7405d85d7a354b8ba49e2db7c4b027e6cb997c1
Author: Namhyung Kim <namhyung@kernel.org>
Date: Thu Oct 31 13:58:30 2013 +0900

description
===========
The cpumode and level in struct addr_localtion was set for a sample
and but updated as cumulative callchains were added.  This led to have
non-matching symbol and cpumode in the output.

Update it accordingly based on the fact whether the map is a part of
the kernel or not.  This is a reverse of what thread__find_addr_map()
does.
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>