[tools] perf/lib/traceevent: Add options to plugins
Message-id: <1410181752-25631-275-git-send-email-jolsa@redhat.com>
Patchwork-id: 91979
O-Subject: [PATCH RHEL7.1 BZ1134356 274/307] tools lib traceevent: Add options to plugins
Bugzilla: 1134356
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1134356
https://bugzilla.redhat.com/show_bug.cgi?id=1134356

upstream
========
commit 5827f2faabe40cc285cc67b697277547a19b6c9a
Author: Steven Rostedt <rostedt@goodmis.org>
Date: Tue Jun 3 18:41:54 2014 -0400

description
===========
The traceevent plugins allows developers to have their events print out
information that is more advanced than what can be achieved by the
trace event format files.

As these plugins are used on the userspace side of the tracing tools, it
is only logical that the tools should be able to produce different types
of output for the events. The types of events still need to be defined by
the plugins thus we need a way to pass information from the tool to the
plugin to specify what type of information to be shown.

Not only does the information need to be passed by the tool to plugin, but
the plugin also requires a way to notify the tool of what options it can
provide.

This builds the plugin option infrastructure that is taken from trace-cmd
that is used to allow plugins to produce different output based on the
options specified by the tool.
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>