[infiniband] mlx4: Avoid null pointer dereference in mlx4_ib_scan_netdevs()
Message-id: <508e887c16e97f825208b1a4295d79b140961f81.1410425016.git.avadai@redhat.com>
Patchwork-id: 93725
O-Subject: [Patch RHEL7.1 V1 183/197] IB/mlx4: Avoid null pointer dereference in mlx4_ib_scan_netdevs()
Bugzilla: 1107617 1107618 1107619 1107620
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>

From: Moni Shoua <monis@mellanox.com>

When Ethernet netdev is not present for a port (e.g. when the link
layer type of the port is InfiniBand) it's possible to dereference
a null pointer when we do netdevice scanning.

To fix that, we move a section of code that needs to run only when
netdev is present to a proper if () statement.

Fixes: ad4885d2 ('IB/mlx4: Build the port IBoE GID table properly under bonding')
Reported-by: Dan Carpenter <dan.carpenter@oracle.com>
Signed-off-by: Moni Shoua <monis@mellanox.com>
Signed-off-by: Or Gerlitz <ogerlitz@mellanox.com>

Upstream: e381835cf1b8e3b2857277dbc3b77d8c5350f70a
Signed-off-by: Amir Vadai <avadai@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>