[infiniband] cxgb4: Only call CQ completion handler if it is armed
Message-id: <1410475447-94392-97-git-send-email-svemuri@redhat.com>
Patchwork-id: 93861
O-Subject: [RHEL7.1 PATCH BZ 1124947 096/108] RDMA/cxgb4: Only call CQ completion handler if it is armed
Bugzilla: 1124947
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

The function __flush_qp() always calls the ULP's CQ completion handler
functions even if the CQ was not armed.  This can crash the system if
the function pointer is NULL. The iSER ULP behaves this way: no
completion handler and never arm the CQ for notification.  So now we
track whether the CQ is armed at flush time and only call the
completion handlers if their CQs were armed.

Also, if the RCQ and SCQ are the same CQ, the completion handler is
getting called twice.  It should only be called once after all SQ and
RQ WRs are flushed from the QP.  So rearrange the logic to fix this.

Signed-off-by: Steve Wise <swise@opengridcomputing.com>
Signed-off-by: Roland Dreier <roland@purestorage.com>
(cherry picked from commit 678ea9b5baab6800692b249bdba77c3c07261d61)
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>