[x86] module: work around kabi module breakage when 16K stacks are enabled
Message-id: <20140912174656.GW29420@redacted.bos.redhat.com>
Patchwork-id: 94058
O-Subject: [RHEL7.1 PATCH 6/6 v4] kabi, x86_64: work around module breakage when 16K stacks are enabled
Bugzilla: 1108378
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Johannes Weiner <jweiner@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Changing the kernel stack size alters some fundamental assumptions
accessing current_thread_info() in modules, since the thread_info struct
is now at a different offset from the kernel stack base than it
previously was found at.

To work around this, we keep a shadow copy of the kernel_stack symbol,
fixed up to contain what would be the expected base address by modules
compiled with the 8K stack size. Then, when an older module is loaded,
we check to see if the symbol being referenced is `kernel_stack' and if
so, fix up the relocation to reference the shadow 8K copy, instead of
the 16K default.

Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>

---

I'm so sorry for this, I hope someone someday outdoes me for ugliest
hack ever in RHEL.

Signed-off-by: Jarod Wilson <jarod@redhat.com>