[powerpc] move debug registers in a structure
Message-id: <1410545655-205645-146-git-send-email-dzickus@redhat.com>
Patchwork-id: 94247
O-Subject: [RHEL7 PATCH 145/626] powerpc: move debug registers in a structure
Bugzilla: 1127366
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Steve Best <sbest@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: David Gibson <dgibson@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=1127366

commit 95791988fec645d196e746fcc0e329e19f7b1347
Author: Bharat Bhushan <r65777@freescale.com>
Date:   Wed Jun 26 11:12:22 2013 +0530

    powerpc: move debug registers in a structure

    This way we can use same data type struct with KVM and
    also help in using other debug related function.

    Signed-off-by: Bharat Bhushan <bharat.bhushan@freescale.com>
    Signed-off-by: Alexander Graf <agraf@suse.de>

Backport notes: due to conflicts with three previously applied patches
code had to move around compare to upstream.  This is what happens
when we backport out of order.

kabi notes:  This patch only changes how the elements in the struct are
accessed.  It does not change the size or layout of the thread_struct.
The reason is the code movement to the debug struct is just a copy and
paste but the new struct is embedded in thread_struct in the same exact
way it was originally.  So GENKSYMS works fine here.  Nothing is different,
other than how things are accessed.

----
v2 - add kabi notes to processor.h
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>