[powerpc] Fix up TOC. for modules
Message-id: <1410545655-205645-432-git-send-email-dzickus@redhat.com>
Patchwork-id: 94560
O-Subject: [RHEL7 PATCH 431/626] powerpc: Fix up TOC. for modules.
Bugzilla: 1127366
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Steve Best <sbest@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: David Gibson <dgibson@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=1127366

commit 4edebbeae3085e71f75584b6582495459e2e6cb2
Author: Rusty Russell <rusty@rustcorp.com.au>
Date:   Tue Mar 18 19:59:26 2014 +1030

    powerpc: Fix up TOC. for modules.

    The kernel resolved the '.TOC.' to a fake symbol, so we need to fix it up
    to point to our .toc section plus 0x8000.

    Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>

kabi reason:  a boolean byte was added just for powerpc.  Of course this
struct is embedded all the way to struct module.  I found a hole at the
end of the struct and used a byte from there to add the extra field.

Therefore the kabi impact should be low.
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>